Generate Classical client and root index
