+++ packages/typespec-ts/src/index.ts
@@ -43,6 +43,7 @@ import { buildRootIndex } from "./modular/buildRootIndex.js";
 import { buildModels } from "./modular/emitModels.js";
 import { buildOperationFiles } from "./modular/buildOperations.js";
 import { buildApiIndexFile } from "./modular/buildApiIndex.js";
+import { buildClassicalClient } from "./modular/buildClassicalClient.js";
 // import { emitPackage, emitTsConfig } from "./modular/buildProjectFiles.js";
 
 export async function $onEmit(context: EmitContext) {
@@ -112,10 +113,12 @@ export async function $onEmit(context: EmitContext) {
     for (const client of modularCodeModel.clients) {
       buildSharedTypes(project, srcPath);
       buildClientContext(client, project, srcPath);
-      buildRootIndex(project, srcPath);
       buildModels(modularCodeModel, project, srcPath);
       buildOperationFiles(client, project, srcPath);
       buildApiIndexFile(project, srcPath);
+      buildClassicalClient(client, project, srcPath);
+      buildRootIndex(client, project, srcPath);
+
       // emitPackage(project, srcPath, modularCodeModel);
       // emitTsConfig(project, srcPath, modularCodeModel);
     }
