+++ packages/typespec-ts/src/modular/helpers/namingHelpers.ts
@@ -1,0 +1,5 @@
+import { Client } from "../modularCodeModel.js";
+
+export function getClientName(client: Client) {
+  return client.name.replace(/Client$/, "");
+}
