+++ packages/typespec-ts/src/modular/helpers/typeHelpers.ts
@@ -63,3 +63,19 @@ export function getType(type: Type): TypeMetadata {
       throw new Error(`Unsupported type ${type.type}`);
   }
 }
+
+export function buildParameterType(
+  clientName: string | undefined,
+  type: Type | undefined
+) {
+  if (!type) {
+    throw new Error("Type should be defined");
+  }
+
+  const typeMetadata = getType(type);
+  let typeName = typeMetadata.name;
+  if (typeMetadata.modifier === "Array") {
+    typeName = `${typeName}[]`;
+  }
+  return { name: clientName ?? "", type: typeName };
+}
